bf.rc: fix bugs
The addition of -- to ascii invocations is pretty self-explanatory; it
would treat programs and input that start with - as options before this.

The program used to set a variable to the former index of the stack on
each [ that isn't skipped. But it would cause problems if you needed to
jump to two ['s in a row, since it's overwritten every time. This fixes
the troubles with programs that have complex bracket usage.